Added final law involving adding ten points based on time
If the time is between 14:01 and 15:59 inclusive, 10 points are
added, otherwise not.

It's assumed that that range is what was intended but it could
be wrong. If the intended range is actually just checking if
the hour is 15, then some receipts will gain 10 more points
than they should. This is a necessary assumption made based on
instructions and the inability to clarify stuff.